chore: some refactoring of postgres code
Move a few functions around for consistency, renamed files.  No actual functionality changes.
Some logging output would produce slightly different results because of
the module name changes.
