+++ martin/src/pg/config_function.rs
@@ -83,3 +83,15 @@ impl PgInfo for FunctionInfo {
         patch_json(tilejson, self.tilejson.as_ref())
     }
 }
+
+impl FunctionInfo {
+    /// For a given function info discovered from the database, append the configuration info provided by the user
+    #[must_use]
+    pub fn append_cfg_info(&self, cfg_inf: &FunctionInfo) -> FunctionInfo {
+        FunctionInfo {
+            // TileJson does not need to be merged because it cannot be de-serialized from config
+            tilejson: self.tilejson.clone(),
+            ..cfg_inf.clone()
+        }
+    }
+}
