+++ martin/src/pg/config_table.rs
@@ -1,11 +1,12 @@
 use std::collections::{BTreeMap, HashMap};
 
+use log::{info, warn};
 use serde::{Deserialize, Serialize};
 use tilejson::{Bounds, TileJSON, VectorLayer};
 
 use crate::config::UnrecognizedValues;
 use crate::pg::config::PgInfo;
-use crate::pg::utils::{patch_json, InfoMap};
+use crate::pg::utils::{normalize_key, patch_json, InfoMap};
 
 pub type TableInfoSources = InfoMap<TableInfo>;
 
@@ -101,5 +102,88 @@ impl PgInfo for TableInfo {
         };
         tilejson.vector_layers = Some(vec![layer]);
         patch_json(tilejson, self.tilejson.as_ref())
+    }
+}
+
+impl TableInfo {
+    /// For a given table info discovered from the database, append the configuration info provided by the user
+    #[must_use]
+    pub fn append_cfg_info(
+        &self,
+        cfg_inf: &TableInfo,
+        new_id: &String,
+        default_srid: Option<i32>,
+    ) -> Option<Self> {
+        // Assume cfg_inf and self have the same schema/table/geometry_column
+        let mut inf = TableInfo {
+            // These values must match the database exactly
+            schema: self.schema.clone(),
+            table: self.table.clone(),
+            geometry_column: self.geometry_column.clone(),
+            // These values are not serialized, so copy auto-detected values from the database
+            geometry_index: self.geometry_index,
+            is_view: self.is_view,
+            tilejson: self.tilejson.clone(),
+            // Srid requires some logic
+            srid: self.calc_srid(new_id, cfg_inf.srid, default_srid)?,
+            prop_mapping: HashMap::new(),
+            ..cfg_inf.clone()
+        };
+
+        match (&self.geometry_type, &cfg_inf.geometry_type) {
+            (Some(src), Some(cfg)) if src != cfg => {
+                warn!(
+                    r#"Table {} has geometry type={src}, but source {new_id} has {cfg}"#,
+                    self.format_id()
+                );
+            }
+            _ => {}
+        }
+
+        let empty = BTreeMap::new();
+        let props = self.properties.as_ref().unwrap_or(&empty);
+
+        if let Some(id_column) = &cfg_inf.id_column {
+            let prop = normalize_key(props, id_column.as_str(), "id_column", new_id)?;
+            inf.prop_mapping.insert(id_column.clone(), prop);
+        }
+
+        if let Some(p) = &cfg_inf.properties {
+            for key in p.keys() {
+                let prop = normalize_key(props, key.as_str(), "property", new_id)?;
+                inf.prop_mapping.insert(key.clone(), prop);
+            }
+        }
+
+        Some(inf)
+    }
+
+    /// Determine the SRID value to use for a table, or None if unknown, assuming self is a table info from the database
+    #[must_use]
+    pub fn calc_srid(&self, new_id: &str, cfg_srid: i32, default_srid: Option<i32>) -> Option<i32> {
+        match (self.srid, cfg_srid, default_srid) {
+            (0, 0, Some(default_srid)) => {
+                info!(
+                    "Table {} has SRID=0, using provided default SRID={default_srid}",
+                    self.format_id()
+                );
+                Some(default_srid)
+            }
+            (0, 0, None) => {
+                let info = "To use this table source, set default or specify this table SRID in the config file, or set the default SRID with  --default-srid=...";
+                warn!("Table {} has SRID=0, skipping. {info}", self.format_id());
+                None
+            }
+            (0, cfg, _) => Some(cfg), // Use the configured SRID
+            (src, 0, _) => Some(src), // Use the source SRID
+            (src, cfg, _) if src != cfg => {
+                warn!(
+                    "Table {} has SRID={src}, but source {new_id} has SRID={cfg}",
+                    self.format_id()
+                );
+                None
+            }
+            (_, cfg, _) => Some(cfg),
+        }
     }
 }
