+++ martin/src/pg/mod.rs
@@ -1,12 +1,12 @@
+mod builder;
 mod config;
 mod config_function;
 mod config_table;
-mod configurator;
 mod errors;
-mod function_source;
 mod pg_source;
 mod pool;
-mod table_source;
+mod query_functions;
+mod query_tables;
 mod tls;
 mod utils;
 
@@ -14,5 +14,5 @@ pub use config::{PgCfgPublish, PgCfgPublishFuncs, PgCfgPublishTables, PgConfig, 
 pub use config_function::FunctionInfo;
 pub use config_table::TableInfo;
 pub use errors::{PgError, PgResult};
-pub use function_source::query_available_function;
 pub use pool::{PgPool, POOL_SIZE_DEFAULT};
+pub use query_functions::query_available_function;
